feat: support pip
